+++ inlang/packages/paraglide/paraglide-js/examples/vite/src/main.ts
@@ -1,4 +1,4 @@
-import * as m from "./paraglide/messages.js";
+import { m } from "./paraglide/messages.js";
 
 // This should be removed by the compiler
 // (can be verified by inspecting the output in dist)