+++ inlang/packages/paraglide/paraglide-next/src/runtime.d.ts
@@ -2,13 +2,13 @@ import type { Runtime } from "@inlang/paraglide-js";
 
 export const baseLocale: Runtime["baseLocale"];
 export const locales: Runtime["locales"];
+export const strategy: Runtime["strategy"];
 export const assertIsLocale: Runtime["assertIsLocale"];
 export const getLocale: Runtime["getLocale"];
 export const setLocale: Runtime["setLocale"];
-export const defineGetLocale: Runtime["defineGetLocale"];
-export const defineSetLocale: Runtime["defineSetLocale"];
+export const overwriteGetLocale: Runtime["overwriteGetLocale"];
+export const overwriteSetLocale: Runtime["overwriteSetLocale"];
 export const isLocale: Runtime["isLocale"];
-export const localeInPath: Runtime["localeInPath"];
-export const localizePath: Runtime["localizePath"];
-export const deLocalizePath: Runtime["deLocalizePath"];
+export const localizeUrl: Runtime["localizeUrl"];
+export const deLocalizeUrl: Runtime["deLocalizeUrl"];
 export const extractLocaleFromRequest: Runtime["extractLocaleFromRequest"];