+++ inlang/packages/sdk/src/project/newProject.test.ts
@@ -1,7 +1,6 @@
 import { expect, test } from "vitest";
 import { defaultProjectSettings, newProject } from "./newProject.js";
 import { loadProjectInMemory } from "./loadProjectInMemory.js";
-import { validate } from "uuid";
 
 test("it should be possible to provide settings for testing or other purposes", async () => {
 	const project = await loadProjectInMemory({
@@ -26,11 +25,20 @@ test("it should be possible to create a project with default settings", async ()
 	expect(settings).toStrictEqual(defaultProjectSettings);
 });
 
-test("it should have a uuid as project id", async () => {
+// for historical reasons, inlang files introduced a project id
+// before lix'es got their own id. having two ids for the same
+// file is not needed anymore.
+test("it should have the lix id as project id", async () => {
 	const project = await loadProjectInMemory({
 		blob: await newProject(),
 	});
+	const { value: lixId } = await project.lix.db
+		.selectFrom("key_value")
+		.select("value")
+		.where("key", "=", "lix_id")
+		.executeTakeFirstOrThrow();
+
 	const projectId = await project.id.get();
 	expect(projectId).toBeDefined();
-	expect(validate(projectId)).toBe(true);
+	expect(projectId).toBe(lixId);
 });